refactor: split server_test.go into focused test files
Split the 4200-line server_test.go into 7 files organized by endpoint:
- server_test.go: shared helpers + server lifecycle tests (~350 lines)
- server_repos_test.go: repos, branches, registration (~365 lines)
- server_jobs_test.go: job listing + enqueue (~1300 lines)
- server_stream_test.go: SSE streaming (~410 lines)
- server_actions_test.go: cancel, rerun, status, comments (~460 lines)
- server_output_test.go: job output + logs (~580 lines)
- server_ops_test.go: batch, remap, fix, review (~820 lines)

All 44 test functions moved verbatim with no logic changes.
